Perf improvement for adding graphs
Issue #4831

The problem was in the code that propagates FK values from principal values when adding dependents. It was doing expensive scanning with a TODO: Perf comment. :-) The fix is to reuse the look-aside indexes already created for fixup. For large graphs the perf is now orders of magnitude faster.